requirements.txt
flask
google-generativeai
python-docx
